filter the output of etcdv3 queries to avoid false results caused by accidental prefix matches
Etcd api v3 has a flat namespace (no directories). So with the following entries in etcd:
/skydns/local/domain
/skydns/local/domain/sub1
/skydns/local/domain/sub2
/skydns/local/domain2
/skydns/local/domain2/sub1
/skydns/local/domain2/sub2

the prefix search with /skydns/local/domain will return all entries and not just the first three.

the filter will return only those result entries which are either exact matches or contain the (path + '/')
as a prefix.